Initial code upload
This code runs on an Arduino MKR zero to datalog electrocardiogram data.
An instrumentation amplifier (Analog devices AD627) is connected to a
sticky pad over the ventricle and a pad below the right clavicle. A
third pad is connected to ground. The Arduino will process the data to
extract pulse, and record the ECG shape to an SD card. At the same time,
this battery powered device can be carried for a day or so to collect
high resolution data.